Reworked k-d-tree and search algorithm in 3d-kriging
An error in the search for the n closest points with the k-d-tree
resolved: instead for excluding points by distance, points, which are
already found, will excluded via id in further search. Also the search
properties in 3d-kriging are now calculated by weighted average of the
variogram model ellipsoids and not from user input.